requirements.txt
tqdm
ftfy
regex
PyYAML
pynrrd
jsonlines
opencv-python
efficientnet_pytorch
